removed the freeze player aspect and the snake enemies
I have had to remove the snake enemies and the freeze player aspects of the game due to issues being caused within the game, I will be changing the enemies to a simpler method. I have also changed the health system to 1 single health in order to make the game more challenging.